requirements_send1_aeh_debug.txt
azure-eventhub==5.12.0